fix(spatial): preserve addressable canvas offsets
Prepared nearest and bilinear samples truncated byte offsets to u32,
wrapping canvases whose RGBA storage crossed four GiB. Store checked
process-width offsets so admitted large canvases sample exact pixels.

The regression builds a large sampling plan without allocating its canvas
and proves bottom-right offsets remain above the former wire-size ceiling.

+++ crates/hypercolor-core/src/spatial/plan.rs
@@ -35,13 +35,13 @@ impl PreparedZoneSamples {
 
 #[derive(Debug, Clone, Copy)]
 pub struct PreparedNearestSample {
-    pub offset: u32,
+    pub offset: usize,
     pub attenuation: u16,
 }
 
 #[derive(Debug, Clone, Copy)]
 pub struct PreparedBilinearSample {
-    pub offsets: [u32; 4],
+    pub offsets: [usize; 4],
     pub x_upper_weight: u8,
     pub y_upper_weight: u8,
     pub attenuation: u16,
